Collector/sigma_scheduler/app/requirements.txt
sqlalchemy
pandas
pyyaml
requests
fastapi
uvicorn
psycopg2-binary
